Added react-router - wroking dummy link to code page, split problems into its own file
