+++ heat-code/src/App.js
@@ -1,12 +1,7 @@
-// App.js
-// Currently the problems page
+import { Problems } from "./Problems";
+import { CodePage } from "./CodePage";
+import React from "react";
 import {
-  Table,
-  Card,
-  CardBody,
-  CardTitle,
-  CardText,
-  Button,
   Navbar,
   NavbarBrand,
   NavbarToggler,
@@ -15,25 +10,9 @@ import {
   Nav,
   Collapse,
 } from "reactstrap";
-import "bootstrap/dist/css/bootstrap.min.css";
-const WelcomeCard = () => {
-  return (
-    <div>
-      <Card outline body>
-        <CardBody>
-          <CardTitle tag="h5">Welcome to HeatCode☄️</CardTitle>
-          <CardText>
-            We're here to help make your interview grind process more
-            competitive and enjoyable. Select a problem to solve while competing
-            head-to-head with someone.
-          </CardText>
-        </CardBody>
-      </Card>
-    </div>
-  );
-};
+import { BrowserRouter as Router, Routes, Route, Link } from "react-router-dom";
 
-const Navigation = () => {
+export const Navigation = () => {
   return (
     <div>
       <Navbar color="light" expand="md" light>
@@ -42,93 +21,49 @@ const Navigation = () => {
         <Collapse navbar>
           <Nav className="me-auto" navbar>
             <NavItem>
-              <NavLink href="/login/">Login</NavLink>
-            </NavItem>
-            <NavItem>
-              <NavLink href="https://github.com/tshaddix18/HeatCode">
-                GitHub
+              <NavLink tag={Link} to="/code">
+                Code
               </NavLink>
             </NavItem>
+            {/*
+              <NavItem>
+                <NavLink to="/login">Login</NavLink>
+              </NavItem>
+              <NavItem>
+                <NavLink href="https://github.com/tshaddix18/HeatCode">
+                  GitHub
+                </NavLink>
+              </NavItem>
+              */}
           </Nav>
         </Collapse>
       </Navbar>
     </div>
   );
 };
-const Problems = () => {
-  return <Table>
-  <thead>
-    <tr>
-      <th>
-        #
-      </th>
-      <th>
-        Problem
-      </th>
-      <th>
-        Difficulty 
-      </th>
-      <th>
-        Tags
-      </th>
-    </tr>
-  </thead>
-  <tbody>
-    <tr>
-      <th scope="row">
-        1
-      </th>
-      <td>
-        Palindrome number
-      </td>
-      <td>
-        Otto
-      </td>
-      <td>
-        @mdo
-      </td>
-    </tr>
-    <tr>
-      <th scope="row">
-        2
-      </th>
-      <td>
-        Jacob
-      </td>
-      <td>
-        Thornton
-      </td>
-      <td>
-        @fat
-      </td>
-    </tr>
-    <tr>
-      <th scope="row">
-        3
-      </th>
-      <td>
-        Larry
-      </td>
-      <td>
-        the Bird
-      </td>
-      <td>
-        @twitter
-      </td>
-    </tr>
-  </tbody>
-</Table>
-}
-
-function App() {
+export default function App() {
   return (
-    <div className="App">
+    <Router>
       <Navigation />
-      <WelcomeCard />
-      <Problems />
-      <p>©2022 Atomic</p>
-    </div>
+      <div>
+        <nav>
+          <ul>
+            <li>
+              <Link to="/">Home</Link>
+            </li>
+            <li>
+              <Link to="/code">Code</Link>
+            </li>
+          </ul>
+        </nav>
+
+        {/* A <Routes> looks through its children <Route>s and
+            renders the first one that matches the current URL. */}
+        <Routes>
+          <Route path="/code" element={<CodePage />} />
+          <Route path="/" element={<Problems />} />
+        </Routes>
+      </div>
+    </Router>
   );
 }
-
-export default App;
