+++ heat-code/src/CodePage.js
@@ -1,0 +1,4 @@
+import React from "react";
+export const CodePage = () => {
+  return <h1>Code</h1>;
+};
